Switch test to kotlin
